Bins size distribution works in the merged class
